move xlsx serialization to datafiles serializers
