+++ tdrs-backend/tdpservice/parsers/urls.py
@@ -1,11 +1,10 @@
 """Routing for DataFiles."""
 from django.urls import path, include
 from rest_framework.routers import DefaultRouter
-from .views import ParsingErrorViewSet, DataFileSummaryViewSet
+from .views import DataFileSummaryViewSet
 
 router = DefaultRouter()
 
-router.register("parsing_errors", ParsingErrorViewSet)
 router.register("dfs", DataFileSummaryViewSet)
 
 urlpatterns = [
